+++ changelog.md
@@ -9,6 +9,7 @@
  - new: methods `.awaitAnimationEnd()`, `.awaitTransitionEnd()` and `.awaitTransAnimEnd()`
  - new: methods `.addClass()`, `.removeClass()` and `.toggleClass()` for perform actions at the end of all CSS transitions and animations
  - fix: `lib.event` API support `handleEvent`
+ - fix: Promise polyfill behavior (call a resolver when the promise is created)
 
 ##v0.2.0